third_party/WebKit/LayoutTests/media/controls/modern/singletap-on-overlay-closes-overflow-menu.html

<!DOCTYPE html>
<html>
<title>Test that the overflow menu closes when the overlay play button area is tapped.</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../media-controls.js"></script>
<video controls width=400 src="../../content/60_sec_video.webm"></video>
<script>
async_test(t => {
  const video = document.querySelector('video');
  var button = overflowButton(video);
  var menu = overflowMenu(video);
  // Need to add a text track for the overflow menu to appear.
  video.addTextTrack("captions");

  video.onloadedmetadata = t.step_func(waitForOverflowButtonToAppear);

  // Start the test once the overflow button has appeared.
  function waitForOverflowButtonToAppear() {
    // Element coordinates will show zero until ready for clicking.
    if (elementCoordinates(button)[0] > 0) {
      openOverflowMenu();
    } else {
      setTimeout(t.step_func(waitForOverflowButtonToAppear));
    }
  }

  function openOverflowMenu() {
    // The overflow menu should be closed by default.
    assert_false(isControlVisible(menu));

    // Open overflow menu.
    const overflowButtonCoordinates = elementCoordinates(button);
    singleTapAtCoordinates(overflowButtonCoordinates[0],
                           overflowButtonCoordinates[1],
                           t.step_func(tapOnOverlay));
  }

  function tapOnOverlay() {
    // Overflow menu should be visible.
    assert_true(isControlVisible(menu));

    // Single tap in the top right hand corner of overlay.
    const overlayCoordinates = coordinatesOutsideElement(mediaControlsOverlayPlayButton(video));
    singleTapAtCoordinates(overlayCoordinates[0] + 1,
                           overlayCoordinates[1] + 1,
                           t.step_func_done(checkOverflowMenuDisappeared));
  }

  function checkOverflowMenuDisappeared() {
    // The overflow menu should have disappeared.
    assert_false(isControlVisible(menu));
  }
});
</script>
</html>
